+++ packages/plugin-node-surrounding-code/test/surrounding-code.test.js
@@ -20,19 +20,19 @@ describe('plugin: node surrounding code', () => {
     client._setDelivery(client => ({
       sendEvent: (payload) => {
         const evt = payload.events[0]
-        expect(Object.keys(evt.stacktrace[0].code))
+        expect(Object.keys(evt.errors[0].stacktrace[0].code))
           .toEqual(['19', '20', '21', '22', '23', '24', '25'])
-        expect(evt.stacktrace[0].code['22'])
+        expect(evt.errors[0].stacktrace[0].code['22'])
           .toBe('    if (cb) this.on(\'finish\', () => cb(this.output()))')
 
-        expect(Object.keys(evt.stacktrace[1].code))
+        expect(Object.keys(evt.errors[0].stacktrace[1].code))
           .toEqual(['28', '29', '30', '31', '32', '33', '34'])
-        expect(evt.stacktrace[1].code['31'])
+        expect(evt.errors[0].stacktrace[1].code['31'])
           .toBe('      return nextLevelUp()')
 
-        expect(Object.keys(evt.stacktrace[2].code))
+        expect(Object.keys(evt.errors[0].stacktrace[2].code))
           .toEqual(['115', '116', '117', '118', '119', '120', '121'])
-        expect(evt.stacktrace[2].code['118'])
+        expect(evt.errors[0].stacktrace[2].code['118'])
           .toBe('  \'Ķ\': \'k\', \'Ļ\': \'L\', \'Ņ\': \'N\', \'Ū\': \'u\'')
 
         done()
@@ -65,9 +65,9 @@ describe('plugin: node surrounding code', () => {
     client._setDelivery(client => ({
       sendEvent: (payload) => {
         const evt = payload.events[0]
-        expect(evt.stacktrace[0].code).toBeTruthy()
-        expect(evt.stacktrace[1].code).toBeUndefined()
-        expect(evt.stacktrace[2].code).toBeTruthy()
+        expect(evt.errors[0].stacktrace[0].code).toBeTruthy()
+        expect(evt.errors[0].stacktrace[1].code).toBeUndefined()
+        expect(evt.errors[0].stacktrace[2].code).toBeTruthy()
         done()
       },
       sendSession: () => {}
@@ -98,13 +98,13 @@ describe('plugin: node surrounding code', () => {
     client._setDelivery(client => ({
       sendEvent: (payload) => {
         const evt = payload.events[0]
-        expect(evt.stacktrace[0].code).toEqual({
+        expect(evt.errors[0].stacktrace[0].code).toEqual({
           1: '// this is just some arbitrary (but real) javascript for testing, taken from',
           2: '// https://github.com/bengourley/source-map-decoder/',
           3: '',
           4: '//'
         })
-        expect(evt.stacktrace[1].code).toEqual({
+        expect(evt.errors[0].stacktrace[1].code).toEqual({
           42: '  return findFile(root, filename)',
           43: '',
           44: '}',
@@ -140,7 +140,7 @@ describe('plugin: node surrounding code', () => {
       sendEvent: (payload) => {
         const endCount = createReadStreamCount
         expect(endCount - startCount).toBe(1)
-        payload.events[0].stacktrace.forEach(stackframe => {
+        payload.events[0].errors[0].stacktrace.forEach(stackframe => {
           expect(stackframe.code).toEqual({
             1: '// this is just some arbitrary (but real) javascript for testing, taken from',
             2: '// https://github.com/bengourley/source-map-decoder/',
@@ -204,7 +204,7 @@ describe('plugin: node surrounding code', () => {
 
     client._setDelivery(client => ({
       sendEvent: (payload) => {
-        payload.events[0].stacktrace.forEach(stackframe => {
+        payload.events[0].errors[0].stacktrace.forEach(stackframe => {
           Object.keys(stackframe.code).forEach(key => {
             expect(stackframe.code[key].length <= 200).toBe(true)
           })