+++ packages/plugin-strip-query-string/strip-query-string.js
@@ -1,12 +1,15 @@
 /*
  * Remove query strings (and fragments) from stacktraces
  */
-const { map } = require('@bugsnag/core/lib/es-utils')
+const { map, reduce } = require('@bugsnag/core/lib/es-utils')
 
 module.exports = {
   init: (client) => {
     client.addOnError(event => {
-      event.stacktrace = map(event.stacktrace, frame => ({ ...frame, file: strip(frame.file) }))
+      const allFrames = reduce(event.errors, (accum, er) => accum.concat(er.stacktrace), [])
+      map(allFrames, frame => {
+        frame.file = strip(frame.file)
+      })
     })
   }
 }