+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Remove Bugsnag* prefix from internal class names [#679](https://github.com/bugsnag/bugsnag-js/pull/679)
 - Remove `client.app` property [#677](https://github.com/bugsnag/bugsnag-js/pull/677)
 - Rename and make private the `Session` method `trackError()` -> `_track()` [#675](https://github.com/bugsnag/bugsnag-js/pull/675)
+- Update `Event` to support multiple errors [#680](https://github.com/bugsnag/bugsnag-js/pull/680)
 
 ## 6.5.0 (2019-12-16)